+++ test/resources.py
@@ -23,6 +23,7 @@ class TestSettings:
     mqtt_topic_json   = 'mqttkit-1/itest/foo/bar/data.json'
     mqtt_topic_event  = 'mqttkit-1/itest/foo/bar/event.json'
     mqtt_topic_homie = 'mqttkit-1/itest/foo/bar/data/__json__'
+    mqtt_topic_json_legacy = 'mqttkit-1/itest/foo/bar/message-json'
 
     # HTTP channel settings.
     channel_path_data = '/mqttkit-1/itest/foo/bar/data'